Refactored massive form container to render form component individually
